src/requirements.txt
fastapi
pillow
requests
transformers
uvicorn
python-multipart